backend/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
langchain
python-dotenv